Made This Way More Efficient
Ok, so, I was looking to replace my sed with cut.
So, I was testing the query on a small sample, when it did what cut
does, which is run through the entire file and make the replacement on
every line.

I went "Woah! I've been a fool doing this the iterative way, the
functional way is way better"
Kind of functional.

So, I filled in the blanks a little more and ended up with this.

It dumps all installed packages into one file, then it dumps all
packages that are depended upon in the whole system into another.
Then it uses a program called comm I found on my system (It may not be
on any other systems) to grab lines unique to installed.
That is, those packages which are installed but not depended upon.

It makes me super happy.
The elegance of it.

And, as a bonus, it went from taking over half an hour to taking about 2
seconds.

Sweet!

Besides comm, it's pretty standards compliant.
Oh, and apt-cache and dpkg, obviously.